% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hann2d.R
\name{hann2d}
\alias{hann2d}
\title{2-D window generator (Hann/Blackman/none)}
\usage{
hann2d(h, w)
}
\arguments{
\item{h}{Height (rows)}

\item{w}{Width (cols)}

\item{window}{One of "hann","blackman","none"}
}
\value{
numeric matrix window of size h x w
}
\description{
2-D window generator (Hann/Blackman/none)
}
