% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gray_matrix.R
\name{read_gray_matrix}
\alias{read_gray_matrix}
\title{Read a grayscale matrix from an image file}
\usage{
read_gray_matrix(path)
}
\arguments{
\item{path}{Path to an image file readable by {imager}}
}
\value{
numeric matrix
}
\description{
Read a grayscale matrix from an image file
}
